image_uploader/requirements.txt
Django>=3.2,<4.0
djangorestframework>=3.12,<4.0
Pillow
celery>=5.1,<6.0
redis>=3.5,<4.0
channels>=3.0,<4.0
channels-redis>=3.3,<4.0
boto3>=1.18,<2.0
django-celery-results>=2.0,<3.0
